Allow unused values in set_header_labels
If you use set_header_labels with headers for columns A, B and C on a flextable with only columns A and B, it will return an error. This quick-fix removes this error. This can be very useful when using a common header/name dictionary, shared across multiple flextables with different column selection. Tested with `trace(edit=TRUE)` successfully.